align alt text behaviour between gif types
